# Speed up EDISP integration