This PR adresses #862 and related issues

I extracted the edisp.peek() method from http://nbviewer.jupyter.org/github/gammapy/gammapy-extra/blob/master/notebooks/data_iact.ipynb (this is mentioned in #862) to [this script](https://github.com/joleroi/gammapy/blob/2d51c79f6a1e1e29671c4c7e1758ff93a396bfe8/dev/profiling/profile_edisp_int.py).

## Status

I get the following performance using gammapy master

```
[profiling] python profile_edisp_int.py 
/home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/axes/_base.py:1322: UserWarning: aspect is not supported for Axes with xscale=log, yscale=linear
  'yscale=%s' % (xscale, yscale))
/home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/image.py:375: UserWarning: Images are not supported on non-linear axes.
  warnings.warn("Images are not supported on non-linear axes.")
Fri Feb 17 13:44:16 2017    restats

         4439850 function calls (4407717 primitive calls) in 7.240 seconds

   Ordered by: cumulative time
   List reduced from 2472 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    7.240    7.240 <string>:1(<module>)
        1    0.001    0.001    7.240    7.240 /home/kingj/Software/gammapy/gammapy/irf/energy_dispersion.py:886(peek)
        1    0.001    0.001    3.834    3.834 /home/kingj/Software/gammapy/gammapy/irf/energy_dispersion.py:665(to_energy_dispersion)
       73    0.169    0.002    3.830    0.052 /home/kingj/Software/gammapy/gammapy/irf/energy_dispersion.py:705(get_response)
     1112    0.050    0.000    2.145    0.002 /home/kingj/Software/gammapy/gammapy/utils/nddata.py:133(evaluate)
        1    0.000    0.000    1.862    1.862 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/pyplot.py:235(show)
        1    0.000    0.000    1.862    1.862 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/backend_bases.py:148(__call__)
        1    0.000    0.000    1.856    1.856 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/backends/backend_qt5.py:147(mainloop)
        1    1.184    1.184    1.856    1.856 {built-in method exec_}
    29437    0.299    0.000    1.554    0.000 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/astropy-1.3-py2.7-linux-x86_64.egg/astropy/units/quantity.py:354(__array_prepare__)
```

Baseline : 7.2 seconds, bottleneck is ``get_response`` btw. calls to ``nddata.evaluate``


## Update

With the reimplementation in this PR I get (after fixes discussed in this PR)

```
[profiling] python profile_edisp_int.py 
/home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/axes/_base.py:1322: UserWarning: aspect is not supported for Axes with xscale=log, yscale=linear
  'yscale=%s' % (xscale, yscale))
/home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/image.py:375: UserWarning: Images are not supported on non-linear axes.
  warnings.warn("Images are not supported on non-linear axes.")
Fri Feb 17 15:59:32 2017    restats

         1910255 function calls (1880397 primitive calls) in 3.583 seconds

   Ordered by: cumulative time
   List reduced from 2472 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    3.583    3.583 <string>:1(<module>)
        1    0.001    0.001    3.583    3.583 /home/kingj/Software/gammapy/gammapy/irf/energy_dispersion.py:847(peek)
        1    0.000    0.000    1.067    1.067 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/pyplot.py:235(show)
        1    0.000    0.000    1.067    1.067 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/backend_bases.py:148(__call__)
        1    0.000    0.000    1.061    1.061 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/backends/backend_qt5.py:147(mainloop)
        1    0.379    0.379    1.061    1.061 {built-in method exec_}
        1    0.000    0.000    1.007    1.007 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/pyplot.py:1369(tight_layout)
        1    0.000    0.000    1.007    1.007 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/figure.py:1718(tight_layout)
        1    0.000    0.000    1.001    1.001 /home/kingj/Software/miniconda3/envs/headversions/lib/python2.7/site-packages/matplotlib/tight_layout.py:263(get_tight_layout_figure)
        1    0.001    0.001    1.001    1.001 /home/kingj/Software/gammapy/gammapy/irf/energy_dispersion.py:665(to_energy_dispersion)

```

Which runs faster and is limited by matplotlib.

## Correctness

Obviously, the integration need not only to run fast but also give correct results. By just looking at the pictures there is a difference (rightmost panel)

### Master

![peek_master](https://cloud.githubusercontent.com/assets/11678435/23065971/6e964fba-f518-11e6-9fa4-50781e511e24.png)

### This PR

![peek_update](https://cloud.githubusercontent.com/assets/11678435/23065995/8af4810e-f518-11e6-94aa-db30930daef0.png)

I think the difference is that the EDISP is normalized to 1 in the integration method that is in master, whereas the new integration does not do any renormalization. @JouvinLea @registerrier @cdeil - which of the images makes more sense?