@cdeil 2017-02-17
The motivation for the current implementation was that HAP-FR results were incorrect before and needed this integration.

Now @registerrier / @leajouvin - The easiest would be if you put an example HAP-FR EDISP file in `gammapy-extra/datasets` and then we use that as an example to discuss / measure precision?