@cdeil 2017-02-17
I consider this a critical fix for the 0.6 release, so if @registerrier or @leajouvin don't have time to review now, or @joleroi, you don't have time tomorrow, then please make this ready so that it can go in (i.e. adapt reference numbers from tests, or add tests if it's not covered)