@registerrier 2017-02-17
OK no time now. But we will look at that with HAP-Fr.