@cdeil 2017-02-17
I left two small inline comments. Otherwise, 👍  to merge if tests using this pass.
(we still have unrelated fails in master unfortunately)
@registerrier and I can then have a look and try it Monday morning from `master` for HAP-FR.