@joleroi 2017-02-17
Thanks for the review. Tests don't pass locally. See description of this PR. I'll wait for input from @registerrier pr @leajouvin on how to normalize the EDISP properly.