@cdeil 2017-02-17
Concerning EDISP normalisation: we need a method to normalise to integrate to 1 for a given `e_true` bin along the `e_reco` axis, no?

But it shouldn't be in this method, but a separate one for flexiblity.
If needed, that other one could then be called from this method optionally `normalise=True`.
I don't know what the default should be here ... probably @registerrier and @leajouvin thought about that, and the new default behaviour should be the same as before this PR, at least for now.