@cdeil 2017-02-17
I've added HAP-FR IRF files for one run just now:
https://github.com/gammapy/gammapy-extra/commit/29bdb50549f846bc5bb92968706a2ce0f8a8f0be

@joleroi - can you please use these two files for testing of this method?
```
$GAMMAPY_EXTRA/datasets/hess-hap-fr-prod1/fullenclosure/run023037/edisp_2d_023037.fits
$GAMMAPY_EXTRA/datasets/hess-crab4-hd-hap-prod2/run023400-023599/run023523/hess_edisp_2d_023523.fits.gz
```
From my side, the goal is to get this PR in now, for the 0.6 release, and to now change / break things at the last minute. So the goal is that the new method to compute the response roughly does what the old one did (just much faster and with simpler code), i.e. resulting output should be similar.