@registerrier 2017-02-17
As far as I  know the code was using the binning of the stored energy dispersion and integrated it on each redo energy bin.
The normalization was done beforehand. 

I suppose that as written the code can not provide a normalized response.

The issue comes from the fact that the evaluate function does not give a normalized probability. 
I don't think we have a clear definition of what should be stored in the e_disp files. 
It can be dN/dm or dN/dlog(m) or even worse, no?
 So it seems to me difficult to get rid of the initial binning in migra.