@registerrier 2017-02-17
I think we can do something simple with np.digitize and np.cumsum