@cdeil 2017-02-18
> The issue comes from the fact that the evaluate function does not give a normalized probability.
> And it is difficult to do so, because I don't think we have a clear definition of what should be stored in the e_disp files.
> It can be dN/dm or dN/dlog(m) or even worse, no?

In the spec this has been defined and pretty clear for a year, no?
http://gamma-astro-data-formats.readthedocs.io/en/latest/irfs/energy_dispersion/index.html
The EDISP in the file should be normalised and return `dP / dmu` when evaluated at a given point.

Now I never checked if that is the case for the current EDISPs we have from the three chains in HESS or CTA, and if the sampling precision is good enough to rely on this normalisation.

It's problematic here because whatever we do by default, it will give incorrect results for someone:
- if we re-normalise, we produce incorrect results near the safe energy cut even if someone filled EDISP fully correctly following the spec
- if we don't re-normalise, we get incorrect fluxes at all energies for EDIPs that aren't filled in a normalised way, as the spec suggests.

So my suggestion here would be:
1. if possible, keep this PR limited to a better implementation, but keep the behaviour roughly the same
2. open a new one to discuss EDISP normalisation, and start a wiki page with some plots how the EDISPs we currently get are filled (from HESS & CTA)

@joleroi @registerrier - OK?