@joleroi 2017-02-19
>The normalization was done beforehand on the migra array from the files themselves.
> I suppose that, as written, the code can not provide a normalized response.

You are right, in that the norm before was calculated on the native migra binning. However, if the goal is just get get a normalized response, it can be achieved by dividing each reco bin by the sum over the entire axis, no? I added that and now the picture now looks as before . The test values still differ though. But I guess the were just put before and are not necessarily 'correct'.

![peek_update2](https://cloud.githubusercontent.com/assets/11678435/23101861/f627e476-f69d-11e6-853c-7c112965a250.png)
 
> if possible, keep this PR limited to a better implementation, but keep the behaviour roughly the same

Agree, see comment above.

> @joleroi - can you please use these two files for testing of this method?

As said before, I don't want to spend much time on fixing the ``get_response`` method.