@joleroi 2017-02-20
When updating the test OGIP file in ``gammapy-extra`` locally, I get the following differences in the fit results

https://gist.github.com/joleroi/dc6023d58c993518400d0be3eaa2078d

Merge? Or continue discussion about the integration method?