@cdeil 2017-02-20
@joleroi - We discussed this just now.
We'll not merge this now, so that things will work tomorrow for the tutorials.
Then tomorrow or Wednesday, @registerrier will merge this or propose something new.